requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
SQLAlchemy==2.0.23
alembic==1.13.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyJWT==2.8.0
stripe==10.11.0
python-dotenv==1.0.1
Flask-WTF==1.2.1
email-validator==2.2.0
Flask-Cors==4.0.1
Flask-Limiter==3.7.0
Faker==27.0.0
marshmallow==3.21.3
sentry-sdk==1.45.0
pytest==8.3.3
pre-commit==3.8.0
